+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.23.0...HEAD
+[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.23.1...HEAD
+
+## [1.23.1][] - 2021-12-02
+
+[1.23.1]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.23.0...1.23.1
 
 ### Added